Fix : 타입 및 API 세분화.
