+++ src/apis/group/index.ts
@@ -1,0 +1,1 @@
+export { getGroups } from './get-groups';
