--- src/components/layouts/index.ts
@@ -1,3 +1,0 @@
-import Layout from './layout';
-
-export { Layout };
