+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import { Layout } from '@/components/layouts';
+import Layout from '@/components/layouts';
 import { QueryProvider } from '@/providers/QueryProvider';
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
