+++ src/queries/group/_root.group.ts
@@ -1,0 +1,1 @@
+export const ROOT_GROUP = 'group';
