+++ src/queries/group/index.ts
@@ -1,0 +1,1 @@
+export { useGetGroupsQuery } from './use-get-groups-query';
