+++ src/queries/user/_root.membership.ts
@@ -1,0 +1,1 @@
+export const ROOT_MEMBERSHIP = 'membership';
