+++ src/queries/user/_root.user.ts
@@ -1,0 +1,1 @@
+export const ROOT_USER = 'user';
