+++ src/types/base/base.date.d.ts
@@ -1,0 +1,4 @@
+export type TBaseDate = {
+  updatedAt: string;
+  createdAt: string;
+};
