+++ src/types/base/base.query.d.ts
@@ -1,0 +1,4 @@
+export type TBaseQuery = {
+  path: string;
+  config?: Record<string, any>;
+};
