+++ src/types/base/index.ts
@@ -1,0 +1,2 @@
+export type { TBaseDate } from './base.date';
+export type { TBaseQuery } from './base.query';
