+++ src/types/common/common.response.d.ts
@@ -1,0 +1,4 @@
+export type TResponse<TData> = {
+  data?: TData;
+  error?: TError;
+};
