+++ src/types/common/index.ts
@@ -1,0 +1,1 @@
+export type { TResponse } from './common.response';
