+++ src/types/group/groups.d.ts
@@ -1,0 +1,1 @@
+export type TGroups = Array<TGroup>;
