+++ src/types/group/index.ts
@@ -1,0 +1,2 @@
+export type { TGroup } from './group';
+export type { TGroups } from './groups';
