+++ src/types/user/user.membership.d.ts
@@ -1,0 +1,8 @@
+export type TMembership = {
+  role: TUserRole;
+  userImage: string;
+  userEmail: string;
+  userName: string;
+  groupId: number;
+  userId: number;
+};
