+++ src/types/user/user.memberships.d.ts
@@ -1,0 +1,8 @@
+import { TGroup } from '../group';
+import { TMembership } from './user.membership';
+
+export type TMemberships = Array<
+  TMembership & {
+    group: TGroup;
+  }
+>;
