+++ src/types/user/user.role.d.ts
@@ -1,0 +1,1 @@
+export type TUserRole = 'ADMIN' | 'MEMBER';
